==> 2. Data Cleaning and Preprocessing .ipynb <==
### Importing Libraries

import pandas as pd 
import numpy as np
import warnings
warnings.simplefilter(action='ignore', category=FutureWarning)
import json 
import ast 

#pd.set_option('display.max_colwidth', -1)

###  Reading Our Webscraped data 

wildfire_df = pd.read_csv('1.raw_data/teamNJV_wildfire_df.csv')
flood_df = pd.read_csv("1.raw_data/teamNJV_flood.csv")
earthquake_df = pd.read_csv("1.raw_data/teamNJV_earthquake.csv")
hurricane_df = pd.read_csv("1.raw_data/teamNJV_hurricane.csv")
tornado_df = pd.read_csv("1.raw_data/teamNJV_tornado.csv")
blizzard_df = pd.read_csv('1.raw_data/teamNJV_blizzard.csv')

tornado_df.head(3)

wildfire_df.head()

### Other Team's Data
#### We used data from other groups in order to better train our model

#first team's data
fire_1_df = pd.read_csv('1.raw_data/ex_pro_1_fire.csv')
blizzard_1_df = pd.read_csv('1.raw_data/ex_pro_1_blizzard.csv', index_col = "Unnamed: 1" )
earthquake_1_df = pd.read_csv('1.raw_data/ex_pro_1_earthquake.csv', index_col = 'Unnamed: 1')
flood_1_df = pd.read_csv('1.raw_data/ex_pro_1_flood.csv', index_col = 'Unnamed: 1')
hurricane_1_df = pd.read_csv('1.raw_data/ex_pro_1_hurricane.csv', )
tornado_1_df = pd.read_csv('1.raw_data/ex_pro_1_tornado.csv', index_col = 'Unnamed: 1')

#second team's data
group_2_comp_df = pd.read_csv('1.raw_data/ex_pro_2_all_articles_final.csv')

#third teams's data 
flood_3_df = pd.read_csv('1.raw_data/ex_pro_3_floods.csv')
earthquake_3_df = pd.read_csv('1.raw_data/ex_pro_3 earthquake.csv', index_col = "Unnamed: 0")

fire_1_df.head()

### Source Getter
##### A function from our last notebook 

def dictionary_conversion(df):
    x_sources = []
    source_dict = {}
    for x in df['source']:
        x_sources.append(x)
    for y in x_sources: 
        source_dict =(ast.literal_eval(y))
        print(source_dict)
        
        return source_dict



This small but might function converts dictionaried strings back into dictionaries, and extracts the name from the dictionary

# tornado_df['source'] = tornado_df['source'].map(lambda x: ast.literal_eval(x)['name'])

### Snippet Renamer
#### The data source was collected from a different api. The plan is to then merge the same disasters. In order to do that the column names need to be the same. 

fire_1_df.rename(columns = {"snippet": "description", 'headline' : 'title'}, inplace = True )
blizzard_1_df.rename(columns = {"snippet": "description", 'headline' : 'title'}, inplace = True)
earthquake_1_df.rename(columns = {"snippet": "description", 'headline' : 'title'}, inplace = True)
flood_1_df.rename(columns = {"snippet": "description", 'headline' : 'title'}, inplace = True)
hurricane_1_df.rename(columns = {"snippet": "description", 'headline' : 'title'}, inplace = True)
tornado_1_df.rename(columns = {"snippet": "description", "Unnamed: 0":'y_label', 'headline' : 'title' }, inplace = True)


### Cleaning up PublishedAt

flood_3_df.rename(columns = {"publishedAt": "pub_date"}, inplace = True )
earthquake_3_df.rename(columns = {"publishedAt": "pub_date"}, inplace = True )
group_2_comp_df.rename(columns = {"publishedAt": "pub_date"}, inplace = True )
tornado_df.rename(columns = {"publishedAt": "pub_date"}, inplace = True )

### Matching the Style of the Datetime format

fire_1_df['pub_date'] = pd.to_datetime(fire_1_df['pub_date']).apply(lambda x: x.date())
blizzard_1_df['pub_date'] = pd.to_datetime(blizzard_1_df['pub_date']).apply(lambda x: x.date())
earthquake_1_df['pub_date'] = pd.to_datetime(earthquake_1_df['pub_date']).apply(lambda x: x.date())
flood_1_df['pub_date'] = pd.to_datetime(flood_1_df['pub_date']).apply(lambda x: x.date())
hurricane_1_df['pub_date'] = pd.to_datetime(hurricane_1_df['pub_date']).apply(lambda x: x.date())
tornado_1_df['pub_date'] = pd.to_datetime(tornado_1_df['pub_date']).apply(lambda x: x.date())
group_2_comp_df['pub_date'] = pd.to_datetime(group_2_comp_df['pub_date']).apply(lambda x: x.date())
flood_3_df['pub_date'] = pd.to_datetime(flood_3_df['pub_date']).apply(lambda x: x.date())
earthquake_3_df['pub_date'] = pd.to_datetime(earthquake_3_df['pub_date']).apply(lambda x: x.date())

#### We scraped our data separatley, therefore there were some formatting discrepancies. The code below fixes that 

flood_df.rename(columns = {"urlToImage": "photo_url"}, inplace = True)
earthquake_df.rename(columns = {"urlToImage": "photo_url"} , inplace = True)
hurricane_df.rename(columns = {"urlToImage": "photo_url"}, inplace = True)
tornado_df.rename(columns = {"urlToImage": "photo_url"}, inplace = True)
blizzard_df.rename(columns = {"urlToImage": "photo_url"}, inplace = True)
group_2_comp_df.rename(columns = {"urlToImage": "photo_url"}, inplace = True)
tornado_df.rename(columns = {"urlToImage": "photo_url"}, inplace = True)

### Matching the URL for Group 1

fire_1_df.rename(columns = {"web_url": "url"}, inplace = True)
blizzard_1_df.rename(columns = {"web_url": "url"}, inplace = True)
earthquake_1_df.rename(columns = {"web_url": "url"}, inplace = True)
flood_1_df.rename(columns = {"web_url": "url"}, inplace = True)
hurricane_1_df.rename(columns = {"web_url": "url"}, inplace = True)
tornado_1_df.rename(columns = {"web_url": "url"}, inplace = True)

### Matching Source Name for Group 3 

flood_3_df.rename(columns = {"source_name": "source"}, inplace = True)
earthquake_3_df.rename(columns = {"source_name": "source"}, inplace = True)

### Identifying Duplicated Values

print("Tornado Duplicates:")
print(tornado_df.duplicated().sum())
print("Wildfire Duplicates:")
print(wildfire_df.duplicated().sum())
print("Blizzard Duplicates")
print(blizzard_df.duplicated().sum())
print("Hurricane Duplicates:")
print(hurricane_df.duplicated().sum())
print("Earthquake Duplicates:")
print(earthquake_df.duplicated().sum())
print("Group 1 Fire Duplicates:")
print(fire_1_df.duplicated().sum())
print("Group 1 Blizzard Duplicates:")
print(blizzard_1_df.duplicated().sum())
print("Group 1 Earthquake Duplicates:")
print(earthquake_1_df.duplicated().sum())
print("Group 1 Flood Duplicates:")
print(flood_1_df.duplicated().sum())
print("Group 1 Hurricane Duplicates:")
print(hurricane_1_df.duplicated().sum())
print("Group 1 Tornado Duplicates:")
print(tornado_1_df.duplicated().sum())
print("Group 2 Duplicates:")
print(group_2_comp_df.duplicated().sum())
print("Group 3 Flood Duplicates")
print(flood_3_df.duplicated().sum())
print("Group 3 Earthquakes")
print(earthquake_3_df.duplicated().sum())


### Dropping Duplicate Values
#### News API only allows one to webscrape for the past 30 days. Furthermore, some website repost articles from other news sources. Therefore we needed to remove the articles with identical titles


wildfire_df.drop_duplicates(keep = "first", inplace = True)
fire_1_df.drop_duplicates(keep = "first", inplace = True)
blizzard_1_df.drop_duplicates(keep = "first", inplace = True)
earthquake_1_df.drop_duplicates(keep = "first", inplace = True)
hurricane_1_df.drop_duplicates(keep = "first", inplace = True)
tornado_1_df.drop_duplicates(keep = "first", inplace = True)
earthquake_3_df.drop_duplicates(keep = "first", inplace = True)

### Separating Group 2

fire_2_df = group_2_comp_df.loc[group_2_comp_df['keyword'] == 'fires']
earthquake_2_df = group_2_comp_df.loc[group_2_comp_df['keyword'] == 'earthquake']
storm_2_df = group_2_comp_df.loc[group_2_comp_df['keyword'] == 'tropical storm']
blizzard_2_df = group_2_comp_df.loc[group_2_comp_df['keyword'] == 'blizzard']
flood_2_df = group_2_comp_df.loc[group_2_comp_df['keyword'] == 'flood']
tornado_2_df = group_2_comp_df.loc[group_2_comp_df['keyword'] == 'tornado']
hurricane_2_df = group_2_comp_df.loc[group_2_comp_df['keyword'] == 'hurricane']

### Format Matching and Column Adding Option 1: Separate Dataframes 

fire_1_df.drop(columns= ['y_label'], inplace = True)
blizzard_1_df.drop(columns= ['y_label'], inplace = True)
earthquake_1_df.drop(columns= ['y_label'], inplace = True)
flood_1_df.drop(columns= ['y_label'], inplace = True)
hurricane_1_df.drop(columns= ['y_label'], inplace = True)
tornado_1_df.drop(columns= ['y_label'], inplace = True)
flood_3_df.drop(columns = ['flood_disaster_relevance'], inplace = True)
earthquake_3_df.drop(columns = ['earthquake_relevance'], inplace = True)

fire_2_df.drop(columns = ["keyword"], inplace = True)
earthquake_2_df.drop(columns = ["keyword"], inplace = True)
storm_2_df.drop(columns = ["keyword"], inplace = True)
blizzard_2_df.drop(columns = ["keyword"], inplace = True)
flood_2_df.drop(columns = ["keyword"], inplace = True)
tornado_2_df.drop(columns = ["keyword"], inplace = True)
hurricane_2_df.drop(columns = ["keyword"], inplace = True)

#dropping the unnamed column from dataframe
# tornado_df.drop(columns = "Unnamed: 0", inplace = True)

#Inputting nulls for the photo url
fire_1_df['photo_url'] = np.NaN
blizzard_1_df['photo_url'] = np.NaN
earthquake_1_df['photo_url'] = np.NaN
flood_1_df['photo_url'] = np.NaN
hurricane_1_df['photo_url'] = np.NaN
tornado_1_df['photo_url'] = np.NaN
flood_3_df['photo_url'] = np.NaN
earthquake_3_df['photo_url'] = np.NaN
#Inputting nulls for the url column
flood_3_df['url'] = np.NaN
earthquake_3_df['url'] = np.NaN

#inputting null values for authors 
fire_1_df['author'] = np.NaN
blizzard_1_df['author'] = np.NaN
earthquake_1_df['author'] = np.NaN
flood_1_df['author'] = np.NaN
hurricane_1_df['author'] = np.NaN
tornado_1_df['author'] = np.NaN
flood_3_df['author'] = np.NaN
earthquake_3_df['author'] = np.NaN

#inputting null values for content 
fire_1_df['content'] = np.NaN
blizzard_1_df['content'] = np.NaN
earthquake_1_df['content'] = np.NaN
flood_1_df['content'] = np.NaN
hurricane_1_df['content'] = np.NaN
tornado_1_df['content'] = np.NaN

### Combined Notebook 

We are trying to decide whether to separate the disasters into seperate dataframes or combine them into one. This is the appropriate notebook to do it in and it gives us options in the future. 

### Combining Similar Disaster DataFrames 

fire_comb = pd.concat([wildfire_df, fire_1_df, fire_2_df])
blizzard_comb = pd.concat([blizzard_df, blizzard_1_df, blizzard_2_df])
earthquake_comb = pd.concat([earthquake_df, earthquake_1_df, earthquake_2_df, earthquake_3_df])
flood_comb = pd.concat([flood_df, flood_1_df, flood_2_df, flood_3_df])
hurricane_comb = pd.concat([hurricane_df, hurricane_1_df, hurricane_2_df])
tornado_comb = pd.concat([tornado_df, tornado_1_df, tornado_2_df])

### Making the Mega Dataset 

comp_df_2 = group_2_comp_df.drop(columns = "keyword")

mega_data = pd.concat([comp_df_2, fire_comb, blizzard_comb, earthquake_comb,
                       flood_comb, hurricane_comb, tornado_comb])

mega_data.head()

### Removing Mega Data duplicates

mega_data.drop_duplicates(inplace= True)

mega_data.shape

### Converting to CSV's 

fire_comb.to_csv('2.processed_data/fire_comb.csv', index = False)
blizzard_comb.to_csv('2.processed_data/blizzard_comb.csv', index = False)
earthquake_comb.to_csv('2.processed_data/earthquake_comb.csv', index = False)
flood_comb.to_csv('2.processed_data/flood_comb.csv', index = False)
hurricane_comb.to_csv('2.processed_data/hurricane_comb.csv', index = False)
tornado_comb.to_csv('2.processed_data/tornado_comb.csv', index = False)
mega_data.to_csv('2.processed_data/mega_data.csv', index = False)
