==> 4. Search Engine.ipynb <==
### Import the Important Libraries 

import pandas as pd 
import numpy as np 
from rake_nltk import Rake 
from nltk.stem import WordNetLemmatizer
from nltk.stem.porter import PorterStemmer
from nltk.tokenize import RegexpTokenizer
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.metrics.pairwise import pairwise_distances, cosine_similarity, linear_kernel
import gensim

df = pd.read_csv('3.final_data/mega_data.csv')
fire_df = pd.read_csv('3.final_data/fire_comb.csv')
blizz_df = pd.read_csv('3.final_data/blizzard_comb.csv')
flood_df = pd.read_csv('3.final_data/flood_comb.csv')
hurr_df = pd.read_csv('3.final_data/hurricane_comb.csv')
earth_df = pd.read_csv('3.final_data/earthquake_comb.csv')
torn_df = pd.read_csv('3.final_data/tornado_comb.csv')

df.head(2)

### Word2Vec

Word2Vec is a two-layer neural net that processes text. It's input is some text and the output is a list of feature vectors that are similar to that original text. The purpose of using Word2Vec is to boost our search engine's capabilities

1. The user inputs keywords such as (california, wildfire)
2. The function then cleans the text string because the word2vec model can only handle lowercase strings
3. The word2vec algorithim finds the most similar words related to that search term and appends those words along with the original search term to a list.
4. The function then sees how many keywords in our dataframe match the keywords in the keywords in the search term list. 
5. The function outputs the title of the most related articles given a threshold for a specific number of search terms 

model = gensim.models.KeyedVectors.load_word2vec_format('lexvec.enwiki+newscrawl.300d.W.pos.vectors')

def art_test(df):
    search = input("Type disaster then location(ex: fire, california):")
    stripped = [x.strip() for x in search.split(' ')]
    result = [x.lower() for x in stripped]
    print(result)
    word2vec_list = []
    for x in model.most_similar(result):
        term,sim = x 
        word2vec_list.append(term)
    word2vec_list.extend(result)
    print(word2vec_list)
    article_list = []
    threshold = 7
    print(threshold)
    while len(article_list) < 5 and threshold > 3:
        for article in df.loc[df['keywords'].map(lambda x: sum(1 for w in word2vec_list if w in x)) == threshold, 'title']:
            article_list.append(article)
        threshold -= 1
    return article_list if len(article_list) > 0 else print('No Results!')

art_test(df)

### Recomender

# Function that takes in arcicle title as input and outputs most similar articles
def get_rcmd_by_min_sim_score(df_l, min_sim_score, title, indices, cosine_sim):
    # Get the index of the article that matches the title
    idx = indices[title]

    # Get the pairwsie similarity scores of all archicles with this article
    sim_scores = list(enumerate(cosine_sim[idx]))

    # Sort the articles based on the similarity scores
    sim_scores = sorted(sim_scores, key=lambda x: x[1], reverse=True)

    # Get the scores which are >= min_sim_score
    filtered_sim_scores = []
    for sim_score in sim_scores:
        if sim_score[1] >= min_sim_score:
            filtered_sim_scores.append(sim_score)
    
    print(f'Number of found articles for {title}: {len(filtered_sim_scores)}')
    filtered_sim_scores.pop(0)
    print(filtered_sim_scores)

    # Get the article indices
    article_indices = [i[0] for i in filtered_sim_scores]
   
    # Return the top most similar articles
    return df_l['title'].iloc[article_indices]

tfidf = TfidfVectorizer(stop_words='english')
tfidf_matrix = tfidf.fit_transform(df["combined_text"])
cosine_sim_matrix = linear_kernel(tfidf_matrix, tfidf_matrix)
cosine_sim_matrix

indices_title = pd.Series(df.index, index=df['title']).drop_duplicates()

title = "Never Try To Drive Through a Flood"

get_rcmd_by_min_sim_score(df, 0.09, title, indices_title, cosine_sim_matrix)

def recomender(df, min_sim_score, title):
    indices_title = pd.Series(df.index, index=df["title"]).drop_duplicates()
    idx = indices_title[title]
    
    tfidf = TfidfVectorizer(stop_words='english')
    tfidf_matrix = tfidf.fit_transform(df["combined_text"])
    cosine_sim_matrix = linear_kernel(tfidf_matrix, tfidf_matrix)
    
    sim_scores = list(enumerate(cosine_sim_matrix[idx]))
    sim_scores = sorted(sim_scores, key=lambda x: x[1], reverse=True)
    
    filtered_sim_scores = []
    for sim_score in sim_scores:
        if sim_score[1] >= min_sim_score:
            filtered_sim_scores.append(sim_score)
    
    print(f'Number of found articles for {title}: {len(filtered_sim_scores)}')
    
    article_indices = [i[0] for i in filtered_sim_scores]
    return df['title'].iloc[article_indices]

### Wildfire

title = fire_df.iloc[0]["title"]
recom_list = list(recomender(df, 0.09, title))

recom_list = recom_list[:3]
temp_df = df
temp_df["final_list"] = temp_df["title"].map(lambda x: 1 if x in recom_list else 0)
temp_df = temp_df.loc[temp_df["final_list"] == 1]
temp_df.shape

temp_df.to_csv("4.search_home/fire_home.csv")

### Tornado

title = torn_df.iloc[5]["title"]
recom_list = list(recomender(df, 0.09, title))

recom_list = recom_list[:3]
temp_df = df
temp_df["final_list"] = temp_df["title"].map(lambda x: 1 if x in recom_list else 0)
temp_df = temp_df.loc[temp_df["final_list"] == 1]
temp_df.shape

temp_df.to_csv("4.search_home/torn_home.csv")

### Flood

title = flood_df.iloc[0]["title"]
recom_list = list(recomender(df, 0.09, title))

recom_list = recom_list[:3]
temp_df = df
temp_df["final_list"] = temp_df["title"].map(lambda x: 1 if x in recom_list else 0)
temp_df = temp_df.loc[temp_df["final_list"] == 1]
temp_df.shape

temp_df.to_csv("4.search_home/flood_home.csv")

### Earthquake

title = earth_df.iloc[1]["title"]
recom_list = list(recomender(df, 0.09, title))

recom_list = recom_list[:3]
temp_df = df
temp_df["final_list"] = temp_df["title"].map(lambda x: 1 if x in recom_list else 0)
temp_df = temp_df.loc[temp_df["final_list"] == 1]
temp_df.shape

temp_df.to_csv("4.search_home/earth_home.csv")

### Blizzard

There are no recent news for blizzard

### Hurricane

title = hurr_df.iloc[0]["title"]
recom_list = list(recomender(df, 0.09, title))

recom_list = recom_list[:3]
temp_df = df
temp_df["final_list"] = temp_df["title"].map(lambda x: 1 if x in recom_list else 0)
temp_df = temp_df.loc[temp_df["final_list"] == 1]
temp_df.shape

temp_df.to_csv("4.search_home/hurr_home.csv")
